test: tobago tree select
Added a test for the Tobago tc:treeSelect component based on Tobago 6.0.0. A demo can be found here: https://tobago-demo.apache.org/demo-6-snapshot/content/090-tree/01-select/Tree_Select.xhtml
The test is based on this demo.

Tobago 6.0.0 uses jsf.js_next_gen 4.0.1-beta.9, which works for tc:treeSelect. Starting with 4.0.2-beta.1, the request form data contains "page:categoriesTree:3:select: page:categoriesTree:3:select". This leads to the issue that a checkbox can be selected but not deselected.

The test might be not optimal, as it fails on 4.0.1-beta.9.

There is a minor issue where the key "jakarta.faces.ClientWindow" is duplicated.